fix: move github button out of header on mobile
